[DEV-6609] refactoring consumers of redux.account.dataAsOf to use new more concise redux property where all info can be derived using helpers.
WithLatestFy -- Exports a hook (for functional component consumers) &
HOC (for class component consumers) which both
(a) hit api
(b) update redux
(c) return asOfDate, allPeriods, and latestPeriod
by invoking releavnt helpers.

Consumers of Hook and HOC:
- TimeFilters.jsx -- use hook for fy instead of dataAsOf.year()
- AgencyContent -- use hook for calendar date, fy and quarter
- DateNote -- use hook for calendar date
- QuarterPickerWithFY & FYPicker -- use hook for fy and period
- AccountContainer -- use HOC for fy
- AccountTimePeriodContainer -- use HOC for fy
- AccountLandingContainer -- use HOC for fy

Consumers Requring Special Changes:

-- CovidHighlights: refactored to get calendar date from
redux via a helper. Previously was expecting a moment to
be in redux, which no longer exists. Did not make a consumer
because it is just too much for this PR (already too big).
-- ObligatedContainer (agency): refactored conditional expression
in componentDidUpdate to ensure all inputs are defined for the
method the lifecycle is calling.